Implement updateSupplierMaterial mutation
