+++ server/schemas/typeDefs.js
@@ -32,13 +32,17 @@ const typeDefs = gql`
     leadTime: Int
   }
 
+  type Material {
+    _id: ID!
+    name: String!
+  }
+
   type Auth {
     token: ID
     user: User
   }
 
   type Query {
-    user: [User]
     suppliers: [Supplier]
     supplier(_id: ID!): Supplier
     me: User
@@ -51,7 +55,7 @@ const typeDefs = gql`
     updateSupplier(_id: ID!, name: String! email: String! address: String! phone: String!): Supplier
     deleteSupplier(_id: ID!):Supplier
     updateUserMaterial(_id: ID!, name: String! stock:Int! safetyStock:Int! anticipatedDemand:Int!): User
-    updateSupplierMaterial(_id: ID!, materialName: String, cost: Float, leadTime: Int): Supplier
+    updateSupplierMaterial(_id: ID!, materialId: ID!, cost: Float!, leadTime: Int!): Supplier
   }
 `;
